requirements.txt
asgiref==3.9.1
cffi==1.17.1
cryptography==45.0.6
Django==5.2.5
djangorestframework==3.16.1
pycparser==2.22
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
